Init Basic account creation
The init with Vite react, this has a structure based on the created vite defaults.
The defaults included extra files that were scrapped for right now. This will be
updated later based on other developers decisions regarding react frameworks to
use.

The current file to pay attention to is the main.jsx, this will be a template of
how we will structure the account creation page even though it currently is not
fully functioning.