Add nginx proxy
Created a configuration file for nginx that will allow the frontend and backend
to be able to communicate properly with each other. It will need some adjustments
as the current config is for the medium article and not the iteration of the
frontend I have for development on HealthPocket.

There is also an additional Dockerfile for when I add in the docker compose
part of the project.