Add create account functionality
This first edits the login.html. The current verison has no way of
getting to a create account page, so this adds a link to a new page.

The app.py was updated to link this button to the new html page and
allow it to have functionality. When a user enters in a new accounts
information and then clicks the submit button. It will read in the
username and password on the flask side and then write it to a csv file.
Then it will redirect them to the home page as well.

The final part is the create_account.html file. This is simply copying
the login page one to one for account create. We can add new fields as
necessary.