Add username check
Usernames could be reused when creating an account, an update was
required to check all the current usernames first to make sure the new
one doesn't exist already. If it does it prints you back to the create
account page and flashes an error. This will also not write to the
account creation csv.