Update dashboard for Med Compare
This adds a new function to the app.py that will check a users
medications and then check the prices of them out of pocket and returns
a json including that price compared to all the insurances in a database
file.

On the frontend a new tab was added that has medicine comparison. This
will give you a table of all your medications based on the ones
currently in the users medication list. The table is basic and not
stylized at all but functional.